测试admob激励广告

+++ live/general/mob/AdMobModel.h
@@ -10,6 +10,6 @@
 
 @interface AdMobModel : NSObject
 
-@property(assign, nonatomic)double count;
+@property(assign, nonatomic)int count;
 @property(copy, nonatomic)NSString *type;
 @end
